requirements.txt
Django==5.0.3
djangorestframework==3.14.0
drf-spectacular==0.28.0
python-dotenv==1.1.1
pytest==8.4.1
gunicorn==23.0.0



